requirements.txt
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
cffi==2.0.0
click==8.1.8
cryptography==46.0.3
exceptiongroup==1.3.1
fastapi==0.128.0
httpx
frozenlist==1.8.0
h11==0.16.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-kasa==0.7.7
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
yarl==1.22.0
